+++ lib/src/iterable_extensions.dart
@@ -601,6 +601,29 @@ extension IterableExtension<T> on Iterable<T> {
       yield slice;
     }
   }
+
+  /// Returns an iterable of the elements in this iterable that are not present
+  /// in [other].
+  ///
+  /// It's aware about occurrence count and removes things only the number of
+  /// times they occurred in [other] but it's indifferent about order items in
+  /// [other].
+  ///
+  /// It assumes the iterable items have consistent `==` and `hashCode`.
+  Iterable<T> subtract(Iterable<T> other) sync* {
+    var elementsCount =
+        other.groupFoldBy<T, int>(identity, (count, _) => (count ?? 0) + 1);
+    for (var element in this) {
+      var count = elementsCount[element];
+      if (count == null) {
+        yield element;
+      } else if (count == 1) {
+        elementsCount.remove(element);
+      } else {
+        elementsCount[element] = count - 1;
+      }
+    }
+  }
 }
 
 /// Extensions that apply to iterables with a nullable element type.